rename sidechain_repl -> repl
